+++ CrudApp/utils/FilePathsUtility.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CrudApp.utils
+{
+    public static class FilePathsUtility
+    {
+        public static string filePath = "database\\WrittenLines.json";
+    }
+}
+
