--- database/.project
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<projectDescription>
-	<name>crud_console</name>
-	<comment></comment>
-	<projects>
-	</projects>
-	<buildSpec>
-	</buildSpec>
-	<natures>
-		<nature>org.jkiss.dbeaver.DBeaverNature</nature>
-	</natures>
-</projectDescription>
