--- database/WrittenLines.json
@@ -1,2 +1,0 @@
-{"Title":"Jeff Sees It","Author":"Angus","PublishYear":1902}
-{"Title":"My Cat","Author":"James Joyce","PublishYear":1951}
