--- database/WrittenLines.txt
@@ -1,2 +1,0 @@
-stan,death,1992
-grift,mother,1902
