--- utils/FilePathsUtility.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace CrudApp.utils
-{
-    public static class FilePathsUtility
-    {
-        public static string filePath = "database\\WrittenLines.json";
-    }
-}
-
